Move all paste-related code to a separate module
