+++ lib/api/behavior.js
@@ -18,8 +18,6 @@ import {
     INPUT_ENTITY_MAP,
 } from '../api/constants';
 
-import DraftUtils from '../api/DraftUtils';
-
 const { hasCommandModifier, isOptionKeyCommand } = KeyBindingUtil;
 
 // TODO Get rid of this as soon as possible.
@@ -270,45 +268,5 @@ export default {
         });
 
         return customStyleMap;
-    },
-
-    filterPastedContent(
-        editorState,
-        maxListNesting,
-        // TODO Implement.
-        enableHorizontalRule,
-        // TODO Implement.
-        enableLineBreak,
-        blockTypes = [],
-        inlineStyles = [],
-        entityTypes = [],
-    ) {
-        let nextEditorState = editorState;
-        const enabledBlockTypes = blockTypes.map(type => type.type);
-        const enabledEntityTypes = entityTypes.map(type => type.type);
-        const enabledInlineStyles = inlineStyles.map(type => type.type);
-
-        nextEditorState = DraftUtils.normaliseBlockDepth(
-            nextEditorState,
-            maxListNesting,
-        );
-
-        nextEditorState = DraftUtils.normaliseBlockType(
-            nextEditorState,
-            enabledBlockTypes,
-        );
-
-        nextEditorState = DraftUtils.normalizeEntityType(
-            nextEditorState,
-            enabledEntityTypes,
-        );
-
-        // TODO Re-test what happens when pasting a styled element, eg. h2 with bold, or blockquote with italic.
-        nextEditorState = DraftUtils.normaliseInlineStyle(
-            nextEditorState,
-            enabledInlineStyles,
-        );
-
-        return nextEditorState;
     },
 };
