+++ lib/components/DraftailEditor.js
@@ -16,6 +16,7 @@ import {
 import DraftUtils from '../api/DraftUtils';
 import behavior from '../api/behavior';
 import conversion from '../api/conversion';
+import normalize from '../api/normalize';
 
 import Toolbar from '../components/Toolbar';
 import Portal from '../components/Portal';
@@ -201,15 +202,16 @@ class DraftailEditor extends Component {
             entityTypes,
         } = this.props;
         const { editorState } = this.state;
-        const contentState = editorState.getCurrentContent();
-        const nextContentState = nextEditorState.getCurrentContent();
-        const isPaste =
-            nextContentState !== contentState &&
+        const content = editorState.getCurrentContent();
+        const nextContent = nextEditorState.getCurrentContent();
+        const shouldFilterPaste =
+            nextContent !== content &&
+            !stripPastedStyles &&
             nextEditorState.getLastChangeType() === 'insert-fragment';
 
         let filteredEditorState = nextEditorState;
-        if (isPaste && !stripPastedStyles) {
-            filteredEditorState = behavior.filterPastedContent(
+        if (shouldFilterPaste) {
+            filteredEditorState = normalize.filterPastedContent(
                 nextEditorState,
                 maxListNesting,
                 enableHorizontalRule,
